Add build and start scripts to package.json
The script section in package.json has been updated to include "build" and "start" scripts. The "build" script utilizes TypeScript compiler (tsc) to transpile our TypeScript into JavaScript. The "start" script is to run the application from the compiled JavaScript files in the dist directory. These are necessary for the proper setup and execution of the application.